+++ Changelog.md
@@ -3,6 +3,7 @@ RestrictEvents Changelog
 #### v1.0.1
 - Disabled `MacPro7,1` RAM & PCI Expansion Slot UIs
 - Disabled `MacBookAir` memory view restrictions
+- Added CPU brand string patch
 
 #### v1.0.0
 - Initial release